# apps/api/internal/modules/tenant_config/infrastructure/persistence/sqlcgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package tenantcfgdb

import (
	"net/netip"

	"github.com/jackc/pgx/v5/pgtype"
)

type Permission struct {
	ID          pgtype.UUID
	Namespace   string
	Description *string
	Status      string
	CreatedAt   pgtype.Timestamptz
	UpdatedAt   pgtype.Timestamptz
	DeletedAt   pgtype.Timestamptz
	CreatedBy   pgtype.UUID
	UpdatedBy   pgtype.UUID
	DeletedBy   pgtype.UUID
}

type Role struct {
	ID          pgtype.UUID
	Name        string
	Description *string
	IsSystem    bool
	Status      string
	CreatedAt   pgtype.Timestamptz
	UpdatedAt   pgtype.Timestamptz
	DeletedAt   pgtype.Timestamptz
	CreatedBy   pgtype.UUID
	UpdatedBy   pgtype.UUID
	DeletedBy   pgtype.UUID
	Version     int32
}

type RolePermission struct {
	RoleID       pgtype.UUID
	PermissionID pgtype.UUID
	GrantedAt    pgtype.Timestamptz
}

type TenantBranding struct {
	ID             pgtype.UUID
	Singleton      bool
	DisplayName    string
	LogoUrl        *string
	PrimaryColor   *string
	SecondaryColor *string
	Timezone       string
	Locale         string
	Status         string
	CreatedAt      pgtype.Timestamptz
	UpdatedAt      pgtype.Timestamptz
	DeletedAt      pgtype.Timestamptz
	CreatedBy      pgtype.UUID
	UpdatedBy      pgtype.UUID
	DeletedBy      pgtype.UUID
	Version        int32
}

type TenantSetting struct {
	ID          pgtype.UUID
	Key         string
	Value       []byte
	Description *string
	Category    *string
	Status      string
	CreatedAt   pgtype.Timestamptz
	UpdatedAt   pgtype.Timestamptz
	DeletedAt   pgtype.Timestamptz
	CreatedBy   pgtype.UUID
	UpdatedBy   pgtype.UUID
	DeletedBy   pgtype.UUID
	Version     int32
}

type User struct {
	ID                  pgtype.UUID
	DocumentType        string
	DocumentNumber      string
	Names               string
	LastNames           string
	Email               *string
	Phone               *string
	PasswordHash        string
	MfaSecret           *string
	MfaEnrolledAt       pgtype.Timestamptz
	FailedLoginAttempts int32
	LockedUntil         pgtype.Timestamptz
	LastLoginAt         pgtype.Timestamptz
	Status              string
	CreatedAt           pgtype.Timestamptz
	UpdatedAt           pgtype.Timestamptz
	DeletedAt           pgtype.Timestamptz
	CreatedBy           pgtype.UUID
	UpdatedBy           pgtype.UUID
	DeletedBy           pgtype.UUID
	Version             int32
}

type UserMfaRecoveryCode struct {
	ID        pgtype.UUID
	UserID    pgtype.UUID
	CodeHash  string
	UsedAt    pgtype.Timestamptz
	Status    string
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
	CreatedBy pgtype.UUID
	UpdatedBy pgtype.UUID
	Version   int32
}

type UserRoleAssignment struct {
	ID               pgtype.UUID
	UserID           pgtype.UUID
	RoleID           pgtype.UUID
	ScopeType        *string
	ScopeID          pgtype.UUID
	GrantedBy        pgtype.UUID
	GrantedAt        pgtype.Timestamptz
	RevokedAt        pgtype.Timestamptz
	RevocationReason *string
	Status           string
	CreatedAt        pgtype.Timestamptz
	UpdatedAt        pgtype.Timestamptz
	CreatedBy        pgtype.UUID
	UpdatedBy        pgtype.UUID
	Version          int32
}

type UserSession struct {
	ID               pgtype.UUID
	UserID           pgtype.UUID
	TokenHash        string
	ParentSessionID  pgtype.UUID
	Ip               *netip.Addr
	UserAgent        *string
	IssuedAt         pgtype.Timestamptz
	ExpiresAt        pgtype.Timestamptz
	RevokedAt        pgtype.Timestamptz
	RevocationReason *string
	Status           string
	CreatedAt        pgtype.Timestamptz
	UpdatedAt        pgtype.Timestamptz
	CreatedBy        pgtype.UUID
	UpdatedBy        pgtype.UUID
	Version          int32
}